Refactored CMMC Compliance stairstep into its own shared component
It is now defined once and used in 3 places in the application.
