+++ CSETWebNg/src/app/assessment/results/mat-cmmc/cmmc-gaps/cmmc-gaps.component.ts
@@ -31,7 +31,7 @@ import {BehaviorSubject} from 'rxjs';
 @Component({
   selector: 'app-cmmc-gaps',
   templateUrl: './cmmc-gaps.component.html',
-  styleUrls: ['../cmmc-results.scss'],
+  styleUrls: ['../../../../../sass/cmmc-results.scss'],
   // tslint:disable-next-line:use-host-property-decorator
   host: { class: 'd-flex flex-column flex-11a' }
 })
