+++ CSETWebNg/src/app/assessment/results/mat-cmmc/cmmc-level-drilldown/cmmc-level-drilldown.component.ts
@@ -29,7 +29,7 @@ import { MaturityService } from '../../../../../app/services/maturity.service';
 @Component({
   selector: 'app-cmmc-level-drilldown',
   templateUrl: './cmmc-level-drilldown.component.html',
-  styleUrls: ['../cmmc-results.scss'],
+  styleUrls: ['../../../../../sass/cmmc-results.scss'],
   // tslint:disable-next-line:use-host-property-decorator
   host: { class: 'd-flex flex-column flex-11a' }
 })
