+++ CSETWebNg/src/app/assessment/results/mat-cmmc/cmmc-level-results/cmmc-level-results.component.ts
@@ -6,7 +6,7 @@ import { MaturityService } from '../../../../../app/services/maturity.service';
 @Component({
   selector: 'app-cmmc-level-results',
   templateUrl: './cmmc-level-results.component.html',
-  styleUrls: ['../cmmc-results.scss'],
+  styleUrls: ['../../../../../sass/cmmc-results.scss'],
   // tslint:disable-next-line:use-host-property-decorator
   host: { class: 'd-flex flex-column flex-11a' },
 })
